Remove use-between & ability to remove photos from GroupPhotos and reflect that in PhotoGallery selection
