+++ package.json
@@ -76,7 +76,6 @@
     "react-spring": "^8.0.27",
     "react-tinder-card": "^1.4.5",
     "realm": "^10.20.0-beta.1",
-    "use-between": "^1.3.4",
     "use-debounce": "^7.0.1"
   },
   "devDependencies": {
